Rename `utils` ➡️  `registry`
